refacotor:NewsCardList컴포넌트 App.jsx에서 분리
